requirements.txt
numpy==1.21.5
scipy==1.4.1
Pillow==9.4.0
cython==0.29.33
matplotlib==3.5.3
scikit-image==0.19.3
tensorflow==2.2.0
keras==2.3.1
protobuf==3.20.3
opencv-python==4.7.0.72
h5py==2.10.0
imgaug==0.4.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.1
json5==0.9.6
notebook==6.5.2
markupsafe==2.0.1
lxml==4.9.1
pandas==1.3.5
Django==3.2.18
djangorestframework==3.14.0
requests==2.28.1
urllib3==1.26.14
IPython